fix: security findings, a migration race, and invented legacy variables
Review of the rename branch. Five things, in descending order of how much
they could cost someone.

freeco_ai_home() renamed a directory, merged its contents and migrated a
database on every call. It reads as an accessor, it has 21 call sites, and
several of them are HTTP route handlers that run concurrently on the async
runtime — so two requests arriving together could both see the legacy
directory and both start moving it. That is a race on the user's only copy
of their data. Resolution and migration now happen once per process.

Removed FRECO_AI_HOME and FRECO_AI_LISTEN. They were read as legacy
variables, but no release ever set either: both are misspellings of the new
names being carried as though they were history. Honouring them means a
typo in someone's Docker config silently works on one machine and not the
next.

Fixed a polynomial-backtracking regex in the JavaScript SDK. `/\/+$/` costs
time quadratic in the length of a trailing run of slashes; the loop that
replaces it is the same length, obviously bounded, and behaves identically
(checked against 200k slashes, 54ms).

Deleted crates/freeco-types/src/naming.rs. It was dead code that nothing
called, and it introduced a third home-directory scheme (`.freeco`)
alongside the two that are real — the worst outcome for a rename whose
whole risk is disagreeing about where the data lives.

Fixed the clippy failures blocking CI: an unnecessary usize cast, and a
collapsible if.

CodeQL's critical "hard-coded cryptographic value used as a nonce" at
freeco-wire/src/peer.rs:1249 is a false positive. The literal is inside
`#[cfg(test)]`, in the test that proves replayed nonces are rejected —
reusing the value is what that test is for. It is left alone.

Which is the other change here: .github/workflows/codeql.yml replaces
default setup, which analyses Rust with build-mode none and so never
compiles the crates. That is why the scan reports only 39% of calls
resolved against a 50% threshold, and why it cannot tell a test literal
from production code. Building the workspace first is what those metrics
measure. Default setup must be turned off in repository settings or the two
configurations conflict and this file is ignored.

The licence files are deliberately untouched: LICENSE-MIT and
LICENSE-APACHE still read "Copyright (c) 2024 OpenFang Contributors", which
is what satisfies the licence and must survive any rename.

+++ crates/freeco-kernel-runtime/src/model_catalog.rs
@@ -4336,7 +4336,7 @@ mod tests {
         for provider in catalog.list_providers() {
             let actual = catalog.models_by_provider(&provider.id).len();
             assert_eq!(
-                provider.model_count as usize, actual,
+                provider.model_count, actual,
                 "{} advertises {} models but {} are registered",
                 provider.id, provider.model_count, actual
             );
